+++ packages/core/src/code_assist/oauth2.ts
@@ -271,7 +271,7 @@ async function initOauthClient(
 
     await triggerPostAuthCallbacks(client.credentials);
   } else {
-    // In Zed integration, we skip the interactive consent and directly open the browser
+    // In ACP mode, we skip the interactive consent and directly open the browser
     if (!config.getAcpMode()) {
       const userConsent = await getConsentForOauth('');
       if (!userConsent) {
